% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatBal.R
\name{WatBal_inVSout}
\alias{WatBal_inVSout}
\title{Annual plot of input vs output}
\usage{
WatBal_inVSout(TimeLoop)
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Annual input and output information
}
\description{
Visulization of annual input and outputs
}
\examples{
WatBal_inVSout <- plot(x=years, y=value)

}
