% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatBal.R
\name{WatBal_majorcomponents}
\alias{WatBal_majorcomponents}
\title{Major components of the water balance}
\usage{
WatBal_majorcomponents(TimeLoop)
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Info on the water balance
}
\description{
Visulization of major components groups in Input, Output and Snow/Glacier
}
\examples{
WatBal_majorcomponents() <- plot(x=varables, y=value)

}
