% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatBal.R
\name{WatBal_plot}
\alias{WatBal_plot}
\title{Daily water balance of the basin}
\usage{
WatBal_plot(TimeLoop)
}
\arguments{
\item{TimeLoop}{The timeloop file present in output folder of J2K model}
}
\value{
Plot with the daily water balance
}
\description{
More info in the basic of results on J2000 hydrological model
}
\examples{
WatBal_plot() <- plot(x=years, y= daily water storage in basin)

}
